Move KclError into kcl_error crate
