+++ rust/Cargo.toml
@@ -48,6 +48,7 @@ rayon = "1.12.0"
 rmp-serde = "1.3.1"
 serde = { version = "1", features = ["derive"] }
 serde_json = { version = "1" }
+schemars = "0.8.22"
 slog = "2.8.2"
 slog-async = "2.8.0"
 slog-json = "2.6.1"
