+++ rust/kcl-error/src/source_range.rs
@@ -1,5 +1,6 @@
 use std::fmt;
 
+use schemars::JsonSchema;
 use serde::Deserialize;
 use serde::Serialize;
 
@@ -37,7 +38,7 @@ impl std::fmt::Display for ModuleId {
 // Don't use a doc comment for the below since the above goes in the website docs.
 // @see isTopLevelModule() in wasm.ts.
 // TODO we need to handle modules better in the frontend.
-#[derive(Debug, Default, Deserialize, Serialize, PartialEq, Copy, Clone, ts_rs::TS, Hash, Eq)]
+#[derive(Debug, Default, Deserialize, Serialize, PartialEq, Copy, Clone, ts_rs::TS, Hash, Eq, JsonSchema)]
 #[ts(export, type = "[number, number, number]")]
 pub struct SourceRange([usize; 3]);
