+++ rust/kcl-lib/src/parsing/token/mod.rs
@@ -7,12 +7,11 @@ use std::num::NonZeroUsize;
 use std::str::FromStr;
 
 use anyhow::Result;
+use kcl_error::KclErrorDetails;
 use parse_display::Display;
 use serde::Deserialize;
 use serde::Serialize;
-use tokeniser::Input;
 use tower_lsp::lsp_types::SemanticTokenType;
-use winnow::error::ParseError;
 use winnow::stream::ContainsToken;
 use winnow::stream::Stream;
 use winnow::{self};
@@ -589,11 +588,7 @@ impl From<&Token> for SourceRange {
 }
 
 pub fn lex(s: &str, module_id: ModuleId) -> Result<TokenStream, KclError> {
-    tokeniser::lex(s, module_id).map_err(From::from)
-}
-
-impl From<ParseError<Input<'_>, winnow::error::ContextError>> for KclError {
-    fn from(err: ParseError<Input<'_>, winnow::error::ContextError>) -> Self {
+    tokeniser::lex(s, module_id).map_err(|err| {
         let (input, offset): (Vec<char>, usize) = (err.input().chars().collect(), err.offset());
         let module_id = err.input().state.module_id;
 
@@ -603,7 +598,7 @@ impl From<ParseError<Input<'_>, winnow::error::ContextError>> for KclError {
             // This is an offset, not an index, and may point to
             // the end of input (input.len()) on eof errors.
 
-            return KclError::new_lexical(crate::errors::KclErrorDetails::new(
+            return KclError::new_lexical(KclErrorDetails::new(
                 "unexpected EOF while parsing".to_owned(),
                 vec![SourceRange::new(offset, offset, module_id)],
             ));
@@ -614,9 +609,9 @@ impl From<ParseError<Input<'_>, winnow::error::ContextError>> for KclError {
         let bad_token = &input[offset];
         // TODO: Add the Winnow parser context to the error.
         // See https://github.com/KittyCAD/modeling-app/issues/784
-        KclError::new_lexical(crate::errors::KclErrorDetails::new(
+        KclError::new_lexical(KclErrorDetails::new(
             format!("found unknown token '{bad_token}'"),
             vec![SourceRange::new(offset, offset + 1, module_id)],
         ))
-    }
+    })
 }
